Fix pressConsumerKey parameter type to uint16_t for full Consumer HID support

+++ atmega_duck/keyboard.cpp
@@ -246,7 +246,7 @@ namespace keyboard {
         sendConsumer(&prev_consumer_report);
     }
 
-    void pressConsumerKey(uint8_t key) {
+    void pressConsumerKey(uint16_t key) {
         prev_consumer_report.key = key;
         sendConsumer(&prev_consumer_report);
         releaseConsumer();
